Creado OperationServiceImpl
Creado OperationServiceImpl y comenzada su implementación.
Se cambia el tipo de la ID de las operaciones a Long.
Se añade un nuevo parámetro a Operation para controlar el IBAN de destino.
Modificadas las properties para que la BDD sea persistida en memoria.-

+++ src/main/java/com/relentlessbank/clientarea/service/OperationService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface OperationService {
     Operation createOperation(int accountId, Operation operation);
     List<Operation> getAllOperationsByBankAccount(int accountId);
-    void deleteOperation(int id);
+    void deleteOperation(long id);
 }
